backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
google-genai>=1.0.0
pandas==2.1.4
Pillow==10.1.0
python-multipart==0.0.6
requests==2.31.0
pytesseract==0.3.10
pymongo==4.6.1
dnspython==2.4.2
twilio==8.10.0
